# docs/swagger.yaml
definitions:
  internal.ContentEntry:
    properties:
      content:
        type: string
      id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
  internal.ContentItem:
    properties:
      entry:
        $ref: '#/definitions/internal.ContentEntry'
      list:
        $ref: '#/definitions/internal.ContentList'
    type: object
  internal.ContentList:
    properties:
      id:
        type: integer
      items:
        items:
          $ref: '#/definitions/internal.ContentItem'
        type: array
      project_id:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /api/content_entries:
    get:
      consumes:
      - application/json
      description: Retrieve list of content entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.ContentEntry'
            type: array
      tags:
      - content
    post:
      consumes:
      - application/json
      description: Create a new content entry
      parameters:
      - description: Content Entry
        in: body
        name: contentEntry
        required: true
        schema:
          $ref: '#/definitions/internal.ContentEntry'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.ContentEntry'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
  /api/content_entries/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a content entry by ID
      parameters:
      - description: Content Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
    get:
      consumes:
      - application/json
      description: Retrieve a content entry by ID
      parameters:
      - description: Content Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.ContentEntry'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
    put:
      consumes:
      - application/json
      description: Update an existing content entry
      parameters:
      - description: Content Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Content Entry
        in: body
        name: contentEntry
        required: true
        schema:
          $ref: '#/definitions/internal.ContentEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.ContentEntry'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
  /api/content_lists:
    get:
      consumes:
      - application/json
      description: Retrieve list of content lists for a project
      parameters:
      - description: Project ID
        in: query
        name: projectid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal.ContentList'
            type: array
      tags:
      - content
    post:
      consumes:
      - application/json
      description: Create a new content list
      parameters:
      - description: Content List
        in: body
        name: contentList
        required: true
        schema:
          $ref: '#/definitions/internal.ContentList'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal.ContentList'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
  /api/content_lists/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a content list by ID
      parameters:
      - description: Content List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
    get:
      consumes:
      - application/json
      description: Retrieve a content list by ID
      parameters:
      - description: Content List ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.ContentList'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
    put:
      consumes:
      - application/json
      description: Update an existing content list
      parameters:
      - description: Content List ID
        in: path
        name: id
        required: true
        type: integer
      - description: Content List
        in: body
        name: contentList
        required: true
        schema:
          $ref: '#/definitions/internal.ContentList'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal.ContentList'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - content
...
swagger: "2.0"